Minor optimisation for Float64Array.

+++ test/benchmark.js
@@ -1,6 +1,6 @@
 var sort = require("../radixsort").sort();
 
-var X = Float32Array;
+var X = Float64Array;
 var max = 1 << 16;
 var ta = new X(max);
